change location and add package
