+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Sandbox_Compose"
         tools:targetApi="31">
         <activity
-            android:name=".AnimationContent"
+            android:name=".animation.AnimationContent"
             android:exported="false"
             android:label="@string/title_activity_animation_content"
             android:theme="@style/Theme.Sandbox_Compose">
